# Improve Character method performance