--- a/samples/Esprima.Benchmark/FactAttribute.cs
+++ b/samples/Esprima.Benchmark/FactAttribute.cs
@@ -1,0 +1,9 @@
+namespace Esprima.Tests;
+
+/// <summary>
+/// Just a helper for compilation of CharBenchmark.
+/// </summary>
+internal sealed class FactAttribute : Attribute
+{
+
+}