… src/Esprima/Character.Generated.cs
… src/Esprima/Character.cs
… src/Esprima/HexConverter.cs
… test/Esprima.Tests/CharMaskGeneratorTest.cs